data-orchestrator-mcp/requirements.txt
# MCP
mcp>=1.0.0

# HTTP
httpx[http2]>=0.27.0

# Extração
pypdf>=4.0.0
polars>=0.20.0
openpyxl>=3.1.0

# Armazenamento
pyarrow>=15.0.0

# Utilitários
pydantic>=2.6.0
python-dotenv>=1.0.0
structlog>=24.1.0
pyyaml>=6.0.0

# Dev/Test
pytest>=8.0.0
pytest-asyncio>=0.23.0
